More

+++ src/cpus/cpu_mips_coproc.cc
@@ -46,7 +46,7 @@
 #include "timer.h"
 
 
-extern volatile int single_step;
+extern volatile uint64_t single_step;
 
 static const char *cop0_names[] = COP0_NAMES;
 static const char *regnames[] = MIPS_REGISTER_NAMES;
